+++ src/include/LSP/Workspace.hpp
@@ -160,6 +160,8 @@ public:
     std::optional<std::vector<lsp::WorkspaceSymbol>> workspaceSymbol(const lsp::WorkspaceSymbolParams& params);
     std::optional<lsp::SemanticTokens> semanticTokens(const lsp::SemanticTokensParams& params, const LSPCancellationToken& cancellationToken);
 
+    lsp::DocumentOnTypeFormattingResult onTypeFormatting(const lsp::DocumentOnTypeFormattingParams& params);
+
     lsp::BytecodeResult bytecode(const lsp::BytecodeParams& params);
     lsp::CompilerRemarksResult compilerRemarks(const lsp::CompilerRemarksParams& params);
     lsp::CodegenResult codeGen(const lsp::CodegenParams& params);