+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- Added support for automatic quote conversion to backticks when typing `{` inside strings. This can be enabled with the `luau-lsp.format.convertQuotes` configuration option. ([#1298](https://github.com/JohnnyMorganz/luau-lsp/pull/1298))
 - Added command `Luau: Compute CodeGen instructions for file` to emit annotated codegen instructions, similar to the bytecode command. External editors can implement this by using the `luau-lsp/codeGen` request. ([#617](https://github.com/JohnnyMorganz/luau-lsp/issues/617))
 
 ### Fixed